created two files to organize api interfaces and calling our Edamam API
