+++ src/users/service/search_recipes/API_Interface/APICaller.java
@@ -1,4 +1,4 @@
-package users.service.search_recipes;
+package users.service.search_recipes.API_Interface;
 
 public interface APICaller {
     public OutputData execute(String[] urlComponents);
