+++ src/users/service/search_recipes/API_Interface/JsonOutputDataConverter.java
@@ -1,0 +1,5 @@
+package users.service.search_recipes.API_Interface;
+
+public interface JsonOutputDataConverter {
+
+}
